storage: mariaDB implementations (#1425)
* mariadb: prepare and finalize statements

* storage: use a new example for tests

* mariadb: change schema types of iota_transaction table

* mariadb: storage_transaction_store

* mariadb: fix error code

* mariadb: add statements in storage functions

* storage: test_utils

* storage: remove defines

* mariadb: fix constness

* storage: fix test_utils

* storage: test store duplicate

* storage: transaction_exist

* storage: move enums to storage.h

* mariadb: transaction_load

* mariadb: transaction_load_metada

* mariadb: impl all loads

* storage: remove 'and' from function names

* storage: factor tests

* storage: load_essence_consensus test

* mariadb: impl and test update functions

* storage: more load tests

* storage: rename solid_state functions

* storage: stubs all functions

* storage: test transaction_meta_clear

* mariadb: store milestone

* mariadb: milestone loads

* mariadb: milestone exists

* mariadb: milestone delete

* mariadb: state_delta store/load

* mariadb: milestone clear

* storage: milestone load tests

* sqlite3: fix tests

* mariadb: spent addresses store/exist

* mariadb: transaction count

* mariadb: approvers count

* storage: fix warnings

* mariadb: start/commit/rollback wrappers

* mariadb: spent addresses set store

* mariadb: batch update solid / snapshot_index

* mariadb: factorize end transaction

* storage: transaction_delete

* mariadb: transactions_delete

* storage: test transactions_delete

* storage: transactions_update_snapshot_index test

* storage: transactions_update_solidity test

* storage: remove old unused tests

* mariadb: storage_transaction_load_hashes

* mariadb: storage_transaction_load_hashes_of_milestone_candidates

* mariadb: fix storage_milestone_exist

* mariadb: storage_transaction_load_hashes_of_approvers

* mariadb: storage_bundle_update_validity

* mariadb: SQL compatibility of storage_statement_transaction_find_build

* mariadb: storage_transaction_find

* sqlite3: fix tests

* storage: storage_transaction_load_hashes_of_approvers tests

* mariadb: fix sql files

* mariadb: fix storage_transaction_load_hashes_of_milestone_candidates

* storage: fix exists SQL statements

* mariadb: use 127.0.0.1 instead of localhost

* mariadb: fix valgrind issues

* mariadb: fix sql files

* storage: fix storage_transaction_load_hashes_of_approvers test

* mariadb: fix hash reset when loading hashes

* mariadb: bind_and_store and bind_and_execute helpers

* mariadb: new errors for statements

* mariadb: execute_statement_one_result helper

* mariadb: add SQL setup script

* mariadb: delta BLOB to LONGBLOB

* infra: ci test setup

* mariadb: fix sql files

* storage: fix tests for TE != 3

* storage: fix test of loaded hashes order

* mariadb: bring back if not exists constraints on index

* mariadb: fix tests for 32bits archs

* mariadb: fix column types

* infra: add bazel_skylib

* infra: update rules_iota

* infra: update rules_iota

+++ ciri/consensus/snapshot/local_snapshots/local_snapshots_manager.h
@@ -33,7 +33,7 @@ typedef struct local_snapshots_manager_s {
   cond_handle_t cond_local_snapshots;
   milestone_tracker_t const *mt;
   snapshots_service_t *snapshots_service;
-  size_t last_snapshot_transactions_count;
+  uint64_t last_snapshot_transactions_count;
   pruning_service_t ps;
 } local_snapshots_manager_t;
